requirements.txt
flask
flask-restx
requests
isort
black
Flask-Cors
Flask-Bcrypt
Flask-PyMongo
pyjwt